requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dohq-artifactory==0.5.241
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.4.2
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
tqdm==4.36.1
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
